refactor ProxyMiddleware into separate middleware and env,host proxy services
